refactor(cart): split CartViewModel into pure math + rates cache
Extracts two concerns out of the 526-line viewmodel:

- CartMath.kt — pure functions (evaluateItem, subtotalOf, totalOf,
  convertAmount, applyConvertedStack, resolveCurrency, resolvedPair)
  that took no viewmodel state beyond what they now receive as params.
  Testable in isolation without an AndroidViewModel harness.

- CartRatesCache.kt — owns the four observeForever subscriptions
  (fees, rates, activeExchange, activeBank) and exposes each source
  as both a LiveData and its last-known scalar. Its `clear()` runs
  from `onCleared` so the viewmodel no longer manages the observer
  bookkeeping itself.

Net: CartViewModel drops from 526 → 446 lines and now reads as a
coordinator over `current` + delegated helpers rather than a mixed
grab-bag of persistence, math, and caching.
